+++ mockserver-core/src/main/java/org/mockserver/mock/action/http/HttpResponseTemplateActionHandler.java
@@ -72,7 +72,7 @@ public class HttpResponseTemplateActionHandler {
 
     private MustacheTemplateEngine getMustacheTemplateEngine() {
         if (mustacheTemplateEngine == null) {
-            mustacheTemplateEngine = new MustacheTemplateEngine(mockServerLogger);
+            mustacheTemplateEngine = new MustacheTemplateEngine(mockServerLogger, configuration);
         }
         return mustacheTemplateEngine;
     }